Replaced gameplay with new type.
The player has two entities on the screen. First he can steer a 'player'
(represented by an yellow circle) which moves around in the arena and
tries to avoid the enemies. Second entity is a fixed torret in the middle
of the arena. The torret can shoot and destroy the enemies to save the
player.